PhotoBlogServer/requirements.txt
# 배포용 의존성 명세
# - 유지보수 용이하게 주요 프레임워크 버전을 고정
Django==4.2.25
djangorestframework==3.15.2
Pillow==10.4.0

